# internal/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Centralized MySQL binlog backup service API.",
        "title": "Binlog Server API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/cluster/overview": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Get cluster overview",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.clusterOverview"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/dashboard": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get task dashboard with replication delay overview",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by source host",
                        "name": "host",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by source port",
                        "name": "port",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.dashboardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sources/lookup": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Lookup tasks by source host and port",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MySQL source host",
                        "name": "host",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "MySQL source port",
                        "name": "port",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.sourceLookupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/summary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get task summary counters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by source host",
                        "name": "host",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by source port",
                        "name": "port",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.summaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create or list tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by source host (GET only)",
                        "name": "host",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by source port (GET only)",
                        "name": "port",
                        "in": "query"
                    },
                    {
                        "description": "Task create payload (name:1-255, cluster_key:[a-zA-Z0-9._-], source/start/storage 按字段规则校验)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tasks.Task"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/tasks.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Create or list tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by source host (GET only)",
                        "name": "host",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by source port (GET only)",
                        "name": "port",
                        "in": "query"
                    },
                    {
                        "description": "Task create payload (name:1-255, cluster_key:[a-zA-Z0-9._-], source/start/storage 按字段规则校验)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tasks.Task"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/tasks.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/workers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "List cluster workers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workers list limit (max 200)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.workerItem"
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.clusterOverview": {
            "type": "object",
            "properties": {
                "leased_task_count": {
                    "type": "integer"
                },
                "running_task_count": {
                    "type": "integer"
                },
                "task_count": {
                    "type": "integer"
                },
                "worker_count": {
                    "type": "integer"
                },
                "workers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.workerItem"
                    }
                }
            }
        },
        "api.createTaskRequest": {
            "type": "object",
            "properties": {
                "cluster_key": {
                    "description": "ClusterKey 集群标识（必填；仅允许 [a-zA-Z0-9._-]；禁止 / \\ ..）。",
                    "type": "string"
                },
                "name": {
                    "description": "Name 任务名（trim 后 1-255 字符）。",
                    "type": "string"
                },
                "source": {
                    "description": "Source 源库配置：host/user 必填且不得含空白；port 1-65535；flavor 为空默认 mysql。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/tasks.SourceConfig"
                        }
                    ]
                },
                "start": {
                    "description": "Start 起点策略：LATEST|FILE_POS|GTID。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/tasks.StartConfig"
                        }
                    ]
                },
                "storage": {
                    "description": "Storage 存储策略：retention_days 必须在 1..3650。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/tasks.Storage"
                        }
                    ]
                }
            }
        },
        "api.dashboardResponse": {
            "type": "object",
            "properties": {
                "generated_at": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.sourceOverview"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/api.summaryResponse"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.dashboardTaskItem"
                    }
                },
                "threshold_seconds": {
                    "type": "integer"
                }
            }
        },
        "api.dashboardTaskItem": {
            "type": "object",
            "properties": {
                "replication": {
                    "$ref": "#/definitions/api.taskReplicationResponse"
                },
                "task": {
                    "$ref": "#/definitions/tasks.Task"
                }
            }
        },
        "api.sourceLookupResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "exists": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "task_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.sourceOverview": {
            "type": "object",
            "properties": {
                "abnormal": {
                    "type": "integer"
                },
                "delayed": {
                    "type": "integer"
                },
                "host": {
                    "type": "string"
                },
                "normal": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "running": {
                    "type": "integer"
                },
                "task_count": {
                    "type": "integer"
                }
            }
        },
        "api.summaryResponse": {
            "type": "object",
            "properties": {
                "abnormal": {
                    "type": "integer"
                },
                "delayed": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "normal": {
                    "type": "integer"
                },
                "retry_backoff": {
                    "type": "integer"
                },
                "running": {
                    "type": "integer"
                },
                "stopped": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.taskReplicationResponse": {
            "type": "object",
            "properties": {
                "delay_seconds": {
                    "type": "integer"
                },
                "has_progress": {
                    "type": "boolean"
                },
                "last_error": {
                    "type": "string"
                },
                "last_event_at": {
                    "type": "string"
                },
                "last_event_file": {
                    "type": "string"
                },
                "last_event_pos": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/tasks.State"
                },
                "status": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "threshold_seconds": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "api.workerItem": {
            "type": "object",
            "properties": {
                "has_updated": {
                    "type": "boolean"
                },
                "host": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "leased": {
                    "type": "integer"
                },
                "online": {
                    "type": "boolean"
                },
                "running": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "task_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
        "tasks.SourceConfig": {
            "type": "object",
            "properties": {
                "flavor": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "semi_sync": {
                    "description": "SemiSync=true 时尝试以 semi-sync 协议拉流；主库不支持时会自动降级为异步。",
                    "type": "boolean"
                },
                "server_id": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "tasks.StartConfig": {
            "type": "object",
            "properties": {
                "file": {
                    "type": "string"
                },
                "gtid_set": {
                    "type": "string"
                },
                "mode": {
                    "$ref": "#/definitions/tasks.StartMode"
                },
                "pos": {
                    "type": "integer"
                }
            }
        },
        "tasks.StartMode": {
            "type": "string",
            "enum": [
                "LATEST",
                "FILE_POS",
                "GTID"
            ],
            "x-enum-varnames": [
                "StartModeLatest",
                "StartModeFilePos",
                "StartModeGTID"
            ]
        },
        "tasks.State": {
            "type": "string",
            "enum": [
                "CREATED",
                "STARTING",
                "RUNNING",
                "LEASE_DEGRADED",
                "REBUILDING_FILE",
                "RETRY_BACKOFF",
                "FAILED",
                "STOPPING",
                "STOPPED"
            ],
            "x-enum-varnames": [
                "StateCreated",
                "StateStarting",
                "StateRunning",
                "StateLeaseDegraded",
                "StateRebuildingFile",
                "StateRetryBackoff",
                "StateFailed",
                "StateStopping",
                "StateStopped"
            ]
        },
        "tasks.Storage": {
            "type": "object",
            "properties": {
                "dir": {
                    "type": "string"
                },
                "retention_days": {
                    "type": "integer"
                }
            }
        },
        "tasks.Task": {
            "type": "object",
            "properties": {
                "cluster_key": {
                    "type": "string"
                },
                "epoch": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_worker_id": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "source": {
                    "$ref": "#/definitions/tasks.SourceConfig"
                },
                "start": {
                    "$ref": "#/definitions/tasks.StartConfig"
                },
                "state": {
                    "$ref": "#/definitions/tasks.State"
                },
                "storage": {
                    "$ref": "#/definitions/tasks.Storage"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}